centralize get/set deploy mode functions
