+++ apstra/blueprint/device_allocation.go
@@ -642,72 +642,3 @@ func (o *DeviceAllocation) deviceProfileNodeIdFromDeviceKey(ctx context.Context,
 
 	o.DeviceProfileNodeId = types.StringValue(result.Items[0].DeviceProfile.Id)
 }
-
-func (o *DeviceAllocation) SetNodeDeployMode(ctx context.Context, client *apstra.Client, diags *diag.Diagnostics) {
-	if o.DeployMode.IsNull() {
-		return
-	}
-
-	if o.DeployMode.IsUnknown() {
-		o.GetNodeDeployMode(ctx, client, diags)
-		return
-	}
-
-	var deployMode apstra.NodeDeployMode
-	err := utils.ApiStringerFromFriendlyString(&deployMode, o.DeployMode.ValueString())
-	if err != nil {
-		diags.AddError(fmt.Sprintf("error parsing deploy mode %q", o.DeployMode.ValueString()), err.Error())
-		return
-	}
-
-	type patch struct {
-		Id         string  `json:"id"`
-		DeployMode *string `json:"deploy_mode"`
-	}
-
-	if err != nil {
-		diags.AddError(fmt.Sprintf("error parsing deploy mode %q", o.DeployMode.ValueString()), err.Error())
-		return
-	}
-
-	var stringPtr *string
-	if deployMode == apstra.NodeDeployModeNone {
-		stringPtr = nil
-	} else {
-		s := deployMode.String()
-		stringPtr = &s
-	}
-
-	setDeployMode := patch{
-		Id:         o.NodeId.ValueString(),
-		DeployMode: stringPtr,
-	}
-	deployModeResponse := patch{}
-
-	err = client.PatchNode(ctx, apstra.ObjectId(o.BlueprintId.ValueString()), apstra.ObjectId(o.NodeId.ValueString()), &setDeployMode, &deployModeResponse)
-	if err != nil {
-		diags.AddError("error setting deploy mode", err.Error())
-		return
-	}
-}
-
-func (o *DeviceAllocation) GetNodeDeployMode(ctx context.Context, client *apstra.Client, diags *diag.Diagnostics) {
-	var node struct {
-		Id         string `json:"id"`
-		Type       string `json:"type"`
-		DeployMode string `json:"deploy_mode"`
-	}
-	err := client.GetNode(ctx, apstra.ObjectId(o.BlueprintId.ValueString()), apstra.ObjectId(o.NodeId.ValueString()), &node)
-	if err != nil {
-		diags.AddError("failed to fetch blueprint node", err.Error())
-	}
-
-	var deployMode apstra.NodeDeployMode
-	err = deployMode.FromString(node.DeployMode)
-	if err != nil {
-		diags.AddError(fmt.Sprintf("error parsing deploy mode %q", node.DeployMode), err.Error())
-		return
-	}
-
-	o.DeployMode = types.StringValue(utils.StringersToFriendlyString(deployMode))
-}
